Fix url scope for new_link
In theforeman/foreman@85a3ebb22fbac6167b20635820a4ce7119f11f96 we have enabled the usage of engine for these.
But this brought an issue, that the scope defaults to the Core routes.
So currently we need to use the engine even if the scope would be fine by default.